android: Bereich „Verwaltung" aus der App entfernen
Die Administration läuft künftig vollständig über den MCP-Server und die
Web-Verwaltung. Der Grund, der bisher für die App sprach — nur das Telefon
könne am Blumenkasten stehen und den Standort übernehmen — trägt nicht mehr:
Claude bekommt im Client die Koordinaten des Geräts, und `ort_anlegen` nimmt
lat/lon entgegen.

Weg sind damit der Bildschirm samt ViewModel und Tests, die Verdrahtung in
HomeScreen/MainActivity/AppContainer, das VerwaltungRepository mit seinen
Endpunkten und Eingabe-DTOs sowie der Auswahlmodus der Karte (ein Tipp auf
die freie Fläche lieferte eine Koordinate — toter Code, der aussähe, als
könne die App noch etwas, das sie nicht mehr kann).

Das Backend behält alle Endpunkte: Sie bedienen Web-Verwaltung und MCP.

`planText` zog nach PlaceDetail.kt um — dort wird es weiterhin gebraucht.
