+++ android/app/src/androidTest/java/de/roessing/app/VergabeScreenshotTest.kt
@@ -27,7 +27,6 @@ import de.roessing.app.ui.HomeScreen
 import de.roessing.app.ui.LeaderboardViewModel
 import de.roessing.app.ui.PlacesViewModel
 import de.roessing.app.ui.IdeenViewModel
-import de.roessing.app.ui.VerwaltungViewModel
 import de.roessing.app.ui.ProfileViewModel
 import de.roessing.app.ui.theme.DorfAppTheme
 import java.io.File
@@ -162,7 +161,6 @@ class VergabeScreenshotTest {
                     leaderboardViewModel = LeaderboardViewModel(FakeStats()),
                     profileViewModel = ProfileViewModel(FakeProfile()),
                     ideenViewModel = IdeenViewModel(FakeIdeen()),
-                    verwaltungViewModel = VerwaltungViewModel(FakeVerwaltung()),
                     onLogout = {},
                 )
             }
