+++ android/app/src/main/java/de/roessing/app/MainActivity.kt
@@ -28,7 +28,6 @@ import de.roessing.app.ui.LeaderboardViewModel
 import de.roessing.app.ui.LoginScreen
 import de.roessing.app.ui.PlacesViewModel
 import de.roessing.app.ui.ProfileViewModel
-import de.roessing.app.ui.VerwaltungViewModel
 import de.roessing.app.ui.VeranstaltungenViewModel
 import de.roessing.app.push.PushZiel
 import de.roessing.app.ui.theme.DorfAppTheme
@@ -114,14 +113,12 @@ private fun Root(pushZiel: PushZiel? = null, onPushZielVerbraucht: () -> Unit = 
             val rangVm: LeaderboardViewModel = viewModel(factory = factory)
             val profilVm: ProfileViewModel = viewModel(factory = factory)
             val ideenVm: IdeenViewModel = viewModel(factory = factory)
-            val verwaltungVm: VerwaltungViewModel = viewModel(factory = factory)
             val termineVm: VeranstaltungenViewModel = viewModel(factory = factory)
             HomeScreen(
                 viewModel = vm,
                 leaderboardViewModel = rangVm,
                 profileViewModel = profilVm,
                 ideenViewModel = ideenVm,
-                verwaltungViewModel = verwaltungVm,
                 veranstaltungenViewModel = termineVm,
                 pushZiel = pushZiel,
                 onPushZielVerbraucht = onPushZielVerbraucht,
@@ -154,8 +151,6 @@ private fun viewModelFactory(container: AppContainer) =
             modelClass.isAssignableFrom(IdeenViewModel::class.java) ->
                 IdeenViewModel(container.ideenRepository) as T
 
-            modelClass.isAssignableFrom(VerwaltungViewModel::class.java) ->
-                VerwaltungViewModel(container.verwaltungRepository) as T
             modelClass.isAssignableFrom(VeranstaltungenViewModel::class.java) ->
                 VeranstaltungenViewModel(container.veranstaltungenRepository) as T
