+++ android/app/src/main/res/values/strings.xml
@@ -170,64 +170,10 @@
     <string name="push_permission_yes">Ja, benachrichtigen</string>
     <string name="push_permission_no">Später</string>
     <string name="push_denied">Alles gut — deine Anfragen stehen weiter hier in der App.</string>
-    <!-- Verwaltung: Orte und Aufgaben pflegen (nur mit der Rolle admin) -->
-    <string name="area_admin_title">Verwaltung</string>
-    <string name="area_admin_subtitle">Orte und Aufgaben pflegen</string>
-    <string name="admin_intro">Hier legst du an, was im Dorf zu tun ist — am besten dort, wo es zu tun ist: Standort übernehmen, Aufgabe eintragen, fertig.</string>
-    <string name="admin_new_place">Neuen Ort anlegen</string>
-    <string name="admin_edit_place">Ort bearbeiten</string>
-    <string name="admin_no_places">Noch kein Ort angelegt.</string>
-    <string name="admin_place_name">Name</string>
-    <string name="admin_place_description">Beschreibung (optional)</string>
-    <string name="admin_place_kind">Art des Ortes</string>
-    <string name="admin_place_position">Standort</string>
-    <string name="admin_place_is_active">Ort ist aktiv</string>
-    <string name="admin_place_active">aktiv</string>
-    <string name="admin_place_paused">pausiert</string>
-    <string name="admin_position_set">%1$.5f, %2$.5f</string>
-    <string name="admin_position_none">Noch kein Standort gewählt</string>
-    <string name="admin_use_my_location">Meinen Standort übernehmen</string>
-    <string name="admin_no_location">Dafür braucht die App die Standortfreigabe.</string>
-    <string name="admin_tap_map">Oder auf die Karte tippen, wo der Ort liegt.</string>
-    <string name="place_kind_box">Blumenkasten</string>
-    <string name="place_kind_bed">Beet</string>
-    <string name="place_kind_other">Sonstiges</string>
-    <string name="admin_new_task">Aufgabe hinzufügen</string>
-    <string name="admin_edit_task">Aufgabe bearbeiten</string>
-    <string name="admin_task_kind">Art der Aufgabe</string>
-    <string name="admin_task_title">Bezeichnung (optional)</string>
-    <string name="admin_task_liters">Liter je Durchgang</string>
-    <string name="admin_task_repeat">Wie oft?</string>
-    <string name="admin_task_regular">Regelmäßig</string>
-    <string name="admin_task_once">Einmalig</string>
-    <string name="admin_task_due">Fällig am</string>
-    <string name="admin_task_due_hint">Format: 2026-08-20. Die Ampel wird drei Tage vorher gelb und rot, sobald der Termin verstrichen ist.</string>
-    <string name="admin_task_interval">Intervall (Tage)</string>
-    <string name="admin_task_red">Rot nach (Tage)</string>
-    <string name="admin_task_remove_when_done">Nach dem Erledigen entfernen</string>
-    <string name="admin_task_is_active">Aufgabe ist aktiv</string>
-    <string name="admin_task_active">läuft</string>
-    <string name="admin_task_paused">pausiert</string>
-    <string name="task_kind_watering">Gießen</string>
-    <string name="task_kind_weeding">Jäten</string>
-    <string name="task_kind_other">Sonstiges</string>
     <string name="task_every_days">alle %1$s Tage</string>
     <string name="task_once">Einmalig</string>
     <string name="task_once_done">Erledigt — diese Aufgabe war einmalig.</string>
     <string name="task_once_due">Einmalig — fällig am %1$s</string>
-    <string name="admin_edit">Bearbeiten</string>
-    <string name="admin_delete">Löschen</string>
-    <string name="admin_save">Speichern</string>
-    <string name="admin_delete_title">Wirklich löschen?</string>
-    <string name="admin_delete_place_text">„%1$s“ wird mit allen Aufgaben und der Historie gelöscht. Wer gerade zugesagt hat, bekommt einen Hinweis.</string>
-    <string name="admin_delete_task_text">„%1$s“ wird mit der Historie gelöscht. Wer gerade zugesagt hat, bekommt einen Hinweis.</string>
-    <string name="admin_delete_yes">Ja, löschen</string>
-    <string name="admin_place_saved">Ort gespeichert.</string>
-    <string name="admin_place_deleted">Ort gelöscht.</string>
-    <string name="admin_task_saved">Aufgabe gespeichert.</string>
-    <string name="admin_task_deleted">Aufgabe gelöscht.</string>
-    <string name="admin_task_paused_toast">Aufgabe pausiert. Wer zugesagt hatte, wurde benachrichtigt.</string>
-    <string name="admin_task_resumed_toast">Aufgabe läuft wieder.</string>
     <!-- Veranstaltungen: „Was ist los in Rössing" — Termine von der Website -->
     <string name="area_events_title">Was ist los?</string>
     <string name="area_events_subtitle">Termine im Dorf.</string>
